backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
drf-spectacular==0.27.1
factory-boy==3.3.0
Faker==24.9.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
packaging==24.0
pillow==10.3.0
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
setuptools==69.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.43.0
